Loading template from JavaScrpt
