--- server/moodle/mod/livequiz/templates/create_question.mustache
@@ -1,35 +1,0 @@
-{{!
-    This file is part of Moodle - http://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-}}
-{{!
-    @template mod_livequiz/index_page
-
-    Index Page for the livequiz plugin
-
-    Classes required for JS:
-    * none
-
-    Context variables required for this template:
-    * fill in later
-
-    Example context (json):
-    * fill in later
-
-}}
-<div id="createQuestionPopup" style="display:none;">
-    <p>This is a pop-up window!</p>
-    <button id="SaveQuestionBtn">Save Question</button>
-</div>
