+++ server/moodle/mod/livequiz/templates/index_page.mustache
@@ -32,6 +32,7 @@
 {{#js}}
     require(['mod_livequiz/question_editor'], (module) => module.init());
 {{/js}}
+
 <div class="main-container">
     <h1>{{pagename}}</h1>
     {{#isteacher}}
